WIP. Progressed unit tests.
