+++ neurological_observations/__init__.py
@@ -1,5 +1,6 @@
 from .controllers import *
 from .models import *
-
-# Test classes
+from .report import *
+from .tests.common import neurological_fixtures
 from .tests.common import nh_clinical_test_neurological_common
+from .tests.common import nh_clinical_test_utils
