+++ neurological_observations/report/__init__.py
@@ -1,0 +1,1 @@
+from . import nh_clinical_observation_report
