+++ neurological_observations/tests/__init__.py
@@ -1,4 +1,6 @@
-# common imported in modules main __init__.py
+# tests.common imported in modules main __init__.py
 from .controllers import *
+from .nh_clinical_observation_report import *
 from .nh_clinical_patient_observation_neurological import *
 from .nh_eobs_api import *
+
