+++ neurological_observations/tests/nh_clinical_observation_report/__init__.py
@@ -1,0 +1,2 @@
+from . import test_get_neurological_observations
+from . import test_get_report_data
