+++ neurological_observations/tests/nh_clinical_observation_report/test_get_neurological_observations.py
@@ -5,15 +5,16 @@ from openerp.tests.common import TransactionCase
 from openerp.tools import DEFAULT_SERVER_DATETIME_FORMAT as DTF
 
 
-class TestGetReportData(TransactionCase):
+class TestGetNeurologicalObservations(TransactionCase):
 
     def setUp(self):
-        super(TestGetReportData, self).setUp()
+        super(TestGetNeurologicalObservations, self).setUp()
         self.test_utils = self.env['nh.clinical.test_utils']
         self.test_utils.admit_and_place_patient()
         self.test_utils.copy_instance_variables(self)
 
-        self.report_wizard_model = self.env['nh.clinical.observation_report_wizard']
+        self.report_wizard_model = self.env[
+            'nh.clinical.observation_report_wizard']
         self.report_model = self.env['report.nh.clinical.observation_report']
 
         self.datetime_start = datetime.now() - timedelta(days=2)
@@ -25,7 +26,19 @@ class TestGetReportData(TransactionCase):
         })
         self.report_wizard.spell_id = self.spell.id
 
-    def test_returns_dict_with_neurological_key_with_list_value(self):
-        report_data = self.report_model.get_report_data(self.report_wizard)
-        self.assertTrue('neurological' in report_data)
-        self.assertTrue(type(report_data['neurological']) is list)
+    def test_returns_list(self):
+        neurological_observations = \
+            self.report_model.get_neurological_observations(self.report_wizard)
+        self.assertTrue(type(neurological_observations) is list)
+
+    def test_no_start_or_end_datetime(self):
+        pass
+
+    def test_start_datetime(self):
+        pass
+
+    def test_end_date_time(self):
+        pass
+
+    def test_start_and_end_datetime(self):
+        pass
